fix: cleanup, file restructure, reply stream handling
